fix: satisfy ruff 0.16 defaults and pin lint tooling
ruff 0.16 promoted isort (I001) and RUF046 into the default rule set,
which turned the nightly Validate run red. Sort the import blocks and
drop the redundant int() around round(), which already returns an int.

Pin ruff and mypy in CI so a future default-rule promotion is a
deliberate bump rather than a surprise failure.

+++ custom_components/ecl310/ecl310.py
@@ -238,7 +238,7 @@ class ECL310Device:
         """Write a scaled temperature to a holding register."""
         await self.async_connect()
         assert self._client is not None
-        raw = int(round(value / scale))
+        raw = round(value / scale)
         result = await self._client.write_register(address, raw, device_id=self._slave)
         if result.isError():
             raise ConnectionError(f"Failed to write temperature to register {address}")
